# Multiple ADB performance and functionality fixes