--- a/pwnlib/adb/protocol.py
+++ b/pwnlib/adb/protocol.py
@@ -46,7 +46,9 @@ class Connection(remote):
         # Try to make sure ADB is running if it's on the default host and port.
         if host == context.defaults['adb_host'] \
         and port == context.defaults['adb_port']:
-            process(context.adb + ['start-server'], level='error').wait_for_close()
+            with context.local():
+                context.clear()
+                process(context.adb + ['start-server'], level='error').wait_for_close()
 
         with context.quiet:
             super(Connection, self).__init__(host, port, level=level, *a, **kw)