@zachriggle 2016-12-14
Apparently the `SYNC:WRITE` command silently swallows writing to a directory, and creates a temporary file.

```
>>> adb.listdir('/data/local/tmp')
[]
>>> adb.write('/data/local/tmp', 'hello')
>>> adb.listdir('/data/local/tmp')
['tmpyXj6mS']
```